Move some tests to coresponding __test__ folders
