+++ src/reducers/__test__/editor-preferences.test.js
@@ -1,5 +1,5 @@
-import reducer from './editor-preferences';
-import {SET_ENCODING, SET_LANGUAGE, SET_LINEFEED, SET_THEME} from "../actions/index";
+import reducer from '../editor-preferences';
+import {SET_ENCODING, SET_LANGUAGE, SET_LINEFEED, SET_THEME} from "../../actions/index";
 
 // Using rewire to get not exported members
 const theme = reducer.__get__('theme');
